Refactor DeleteExperimentDialog into separate component
